refactor(cli): group advisor warning notification
