+++ docs/projects/civ7-live-play-support/topics/official-runtime-affordance-inventory.md
@@ -33,7 +33,7 @@ the current tuner state.
 | Notification inventory and blocker identity | `.civ7/outputs/resources/Base/modules/base-standard/ui/notification-train/model-notification-train.js:40`; `:45` | replace raw HUD guessing | extend `game play notifications list`; add `notification-details` | high | ids expire after actions |
 | Display queue manager | `.civ7/outputs/resources/Base/modules/base-standard/ui/display-queue/display-queue-manager.js:154`; `:174`; `:276`; `:293`; `:336` | UI queue/popup ordering | `game play ui-queue` lens over active, pending, suspended displays | high | module singleton may not be globally reachable |
 | Notification handler map | `.civ7/outputs/resources/Base/modules/base-standard/ui/notification-train/notification-handlers.js:188`; `:222`; `:279`; `:369`; `:407`; `:438`; `:539`; `:565` | replace text matching with official handler categories | static `notification-handler-map` catalog feeding HUD hints | high | registration table still needs a mapping audit |
-| Advisor warning closeout | `.civ7/outputs/resources/Base/modules/base-standard/ui/notification-train/notification-handlers.js:585`; `.civ7/outputs/resources/Base/modules/base-standard/data/advisory.xml:13` | avoid raw dismissal for advisor blockers | keep `advisor-warning`; add `advisor-warnings` read lens | high | WatchOut manager can auto-dismiss in some paths |
+| Advisor warning closeout | `.civ7/outputs/resources/Base/modules/base-standard/ui/notification-train/notification-handlers.js:585`; `.civ7/outputs/resources/Base/modules/base-standard/data/advisory.xml:13` | avoid raw dismissal for advisor blockers | keep `notifications advisor-warning`; add `advisor-warnings` read lens | high | WatchOut manager can auto-dismiss in some paths |
 | Production recommendations | `.civ7/outputs/resources/Base/modules/base-standard/ui/production-chooser/production-chooser-helpers.chunk.js:1285` | replace manual build advice heuristics | `game play advisor-builds --city-id ...` | medium-high | only production recommendation is proven here |
 | Diplomacy response data | `.civ7/outputs/resources/Base/modules/base-standard/ui/notification-train/notification-handlers.js:539`; `panel-diplomacy-project-reaction.js:374` | inspect available responses and costs | `game play diplomacy-response --action-id ...` | high | action id is not notification id |
 | Relationship and agenda data | `.civ7/outputs/resources/Base/modules/base-standard/ui/diplomacy/relationship-breakdown.chunk.js:40`; `relationship-tooltip.js:149`; `panel-diplomacy-actions.js:769` | replace report scraping | `game play relations --player-id ...` | high | reports can be informational, not response-required |
