+++ plugins/cli/topics/game/package.json
@@ -37,7 +37,7 @@
         "description": "Validate and send diplomatic action responses and first-meet greetings"
       },
       "game:play:notifications": {
-        "description": "Read, schedule, and dismiss live Civ7 notifications"
+        "description": "Read and schedule live Civ7 notifications; dismiss reviewed items and acknowledge advisor warnings"
       },
       "game:play:screen": {
         "description": "Inspect and dismiss App UI display-queue screens (cinematic moments)"
